+++ src/blocks.js
@@ -11,6 +11,14 @@
 
 import './block/block.js';
 
+/**
+ * Main Gutenberg Forms block
+ */
+import './blocks/gutenberg-forms/index.js';
+
+/**
+ * Inner blocks
+ */
 import './blocks/checkbox/index.js';
 import './blocks/column/index.js';
 import './blocks/datepicker/index.js';
